fix: move rule to common folder
