+++ packages/core/src/rules/common/__tests__/spec-ref-validation.test.ts
