+++ packages/core/src/rules/common/spec-ref-validation.ts
@@ -1,7 +1,7 @@
-import { Oas3Rule } from '../../visitors';
+import { Oas2Rule, Oas3Rule } from '../../visitors';
 import { isRef } from '../../ref-utils';
 
-export const SpecRefValidation: Oas3Rule = () => {
+export const SpecRefValidation: Oas3Rule | Oas2Rule = () => {
   const nodesToValidate = [
     'Schema',
     'Response',
