+++ packages/core/src/rules/oas2/index.ts
@@ -40,6 +40,7 @@ import { ResponseContainsHeader } from '../common/response-contains-header';
 import { ResponseContainsProperty } from './response-contains-property';
 import { ScalarPropertyMissingExample } from '../common/scalar-property-missing-example';
 import { RequiredStringPropertyMissingMinLength } from '../common/required-string-property-missing-min-length';
+import { SpecRefValidation } from '../common/spec-ref-validation';
 
 export const rules = {
   spec: OasSpec as Oas2Rule,
@@ -84,6 +85,7 @@ export const rules = {
   'response-contains-property': ResponseContainsProperty as Oas2Rule,
   'scalar-property-missing-example': ScalarPropertyMissingExample,
   'required-string-property-missing-min-length': RequiredStringPropertyMissingMinLength,
+  'spec-ref-validation': SpecRefValidation,
 };
 
 export const preprocessors = {};
